weather/requirements.txt
pandas
pyarrow
sqlalchemy
psycopg2-binary
tqdm
requests
pyyaml
openmeteo-requests
retry-requests
